+++ CHANGELOG.md
@@ -3,11 +3,25 @@
 * Minor releases (X.1.X) are new features such as added functions or small changes that don't cause major compatibility issues.
 * Major releases (1.X.X) are major new features or changes that break backward compatibility in a big way.
 
-## [Latest](https://github.com/int-brain-lab/iblutil/commits/main) [1.15.0]
+
+## [Latest](https://github.com/int-brain-lab/iblutil/commits/main) [1.17.0]
 
 ### Modified
 - upgrade minimum supported Python version to 3.10
 - io.params.read returns the json file name if a decoding error occurs
+
+## [1.16.0]
+
+### Added
+
+- io.binary.write_array: write array to binary file
+
+## [1.15.0]
+
+### Added
+
+- io.binary.load_as_dataframe: read binary data as a Pandas DataFrame
+- io.binary.convert_to_parquet: convert a binary file to Parquet
 
 ## [1.14.0]